Title: 실수와 문자 받아 출력하기, Time: 176ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240901/실수와 문자 받아 출력하기/print-number--and-letter.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main (String args[]) {
         Scanner sc = new Scanner(System.in);
         char c = sc.next().charAt(0);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
         System.out.printf("%c\n%.2f\n%.2f",c,a,b);
     }
 }
